requirements.txt
pylint
pytest
pika
psycopg2-binary